fix(app): keep a mock factory out of the module it mocks
`ImportModal.options.test.tsx` mocks `@/services/api` with a factory that
reaches for the recorded parse of the fixture it pastes. That helper lived in
`import-preview.testkit.ts`, which imports `apiService` in order to spy on it -
so the factory pulled in a module that imports the module being mocked, and the
module runner does not fail on that cycle: it deadlocks. The file hung, the
worker never reported, and CI killed the whole app job at its timeout with no
output naming a cause (three jobs on #878, ~20 minutes each).

`recordedParse` moves to its own module that imports nothing of the app, so a
mock factory can use it without closing a cycle; the testkit re-exports it for
the cases that are not inside a factory.
